split up RoleEmojiCommand into command+subcommand
