+++ src/main/java/net/javadiscord/javabot/systems/staff_commands/role_emoji/AddRoleEmojiSubcommand.java
@@ -1,45 +1,40 @@
-package net.javadiscord.javabot.systems.staff_commands;
+package net.javadiscord.javabot.systems.staff_commands.role_emoji;
 
 import org.jetbrains.annotations.NotNull;
 
-import com.dynxsty.dih4jda.interactions.commands.SlashCommand;
+import com.dynxsty.dih4jda.interactions.commands.SlashCommand.Subcommand;
 
-import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Message.Attachment;
 import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import net.dv8tion.jda.api.interactions.commands.DefaultMemberPermissions;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
-import net.dv8tion.jda.api.interactions.commands.build.Commands;
-import net.dv8tion.jda.api.interactions.commands.build.SlashCommandData;
+import net.dv8tion.jda.api.interactions.commands.build.SubcommandData;
 import net.javadiscord.javabot.data.config.BotConfig;
 import net.javadiscord.javabot.util.Checks;
 import net.javadiscord.javabot.util.Responses;
 
 /**
- * This class represents the /add-role-emoji command.
- * This command allows adding emojis which are usable only be members with certain roles.
+ * This class represents the /emoji-admin add subcommand.
+ * This subcommand allows adding emojis which are usable only be members with certain roles.
  */
-public class RoleEmojiCommand extends SlashCommand{
+public class AddRoleEmojiSubcommand extends Subcommand{
 
 	private final BotConfig botConfig;
 
 	/**
-	 * The constructor of this class, which sets the corresponding {@link net.dv8tion.jda.api.interactions.commands.build.SlashCommandData}.
+	 * The constructor of this class, which sets the corresponding {@link SubcommandData}s.
 	 * @param botConfig The main configuration of the bot
 	 */
-	public RoleEmojiCommand(BotConfig botConfig) {
+	public AddRoleEmojiSubcommand(BotConfig botConfig) {
 		this.botConfig=botConfig;
-		SlashCommandData slashCommandData = Commands.slash("add-role-emoji", "Adds an emoji only usable only with certain roles")
+		SubcommandData subCommandData = new SubcommandData("add", "Adds an emoji only usable only with certain roles")
 				.addOption(OptionType.STRING, "name", "The name of the emoji", true)
-				.addOption(OptionType.ATTACHMENT, "emoji", "the emoji", true)
-				.setDefaultPermissions(DefaultMemberPermissions.enabledFor(Permission.ADMINISTRATOR))
-				.setGuildOnly(true);
+				.addOption(OptionType.ATTACHMENT, "emoji", "the emoji", true);
 		for (int i = 1; i <= 10; i++) {
-			slashCommandData.addOption(OptionType.ROLE, "role-"+i, "A role allowed to use the emoji", i==1);
+			subCommandData.addOption(OptionType.ROLE, "role-"+i, "A role allowed to use the emoji", i==1);
 		}
-		setSlashCommandData(slashCommandData);
+		setSubcommandData(subCommandData );
 		requireUsers(botConfig.getSystems().getAdminConfig().getAdminUsers());
 	}
